+++ Source/Machine.Fakes.Adapters.Specs/SampleCode/ILookupService.cs
@@ -1,0 +1,7 @@
+﻿namespace Machine.Fakes.Adapters.Specs.SampleCode
+{
+    public interface ILookupService
+    {
+        void TryLookup(string key, out string value); 
+    }
+}