+++ Source/Machine.Fakes.Adapters.Specs/SampleCode/IReturnOutAndRef.cs
@@ -1,0 +1,7 @@
+﻿namespace Machine.Fakes.Adapters.Specs.SampleCode
+{
+    public interface IReturnOutAndRef
+    {
+        void GetTwoValues(string input, out string output, ref object additional);
+    }
+}